+++ example/example.dart
@@ -1,0 +1,22 @@
+import 'dart:math' show Random;
+import 'package:trotter/trotter.dart';
+
+void main() {
+  final rand = Random(),
+      perms = ['Steinhaus', 'Johnson', 'Trotter'].permutations(),
+      dissent = [
+    'No, no, no!',
+    'I disagree.',
+    'That is just wrong!',
+    'Preposterous!'
+  ],
+      exclaim = () => dissent[rand.nextInt(dissent.length)],
+      title = (List<String> names) => names.join('-') + ' ordering';
+
+  print('Gentlemen, I propose we call the ordering of permutations the');
+  print('${title(perms().first)}.');
+  for (final perm in perms().skip(1)) {
+    print("${exclaim()} It should be called the ${title(perm)}!");
+  }
+  print("\n\nOkay then... we'll vote on it!");
+}