+++ lib/src/trotter_helpers.dart
@@ -3,31 +3,31 @@ part of trotter;
 /// Cache to store calculated factorials.
 Map<int, BigInt> _factCache = {};
 
-/// Calculates [n]!
+/// Calculates `n`!
 BigInt _fact(int n) => _factCache.containsKey(n)
     ? _factCache[n]
     : (n < 2 ? BigInt.one : _factCache[n] = BigInt.from(n) * _fact(n - 1));
 
-/// Calculates the number of permutations of [r] items taken from [n].
+/// Calculates the number of permutations of `r` items taken from `n`.
 BigInt _nPr(int n, int r) => _fact(n) ~/ _fact(n - r);
 
-/// Calculates the number of combinations of [r] items taken from [n].
+/// Calculates the number of combinations of `r` items taken from `n`.
 BigInt _nCr(int n, int r) => _nPr(n, r) ~/ _fact(r);
 
-/// Returns the items in [arrangement] in the same order as they appear in [items].
+/// Returns the items in `arrangement` in the same order as they appear in `items`.
 List _sortedArrangement(List arrangement, List items) =>
     (List.from(arrangement))
       ..sort((x, y) => items.indexOf(x).compareTo(items.indexOf(y)));
 
-/// Checks whether the items in [items] are unique.
+/// Checks whether the items in `items` are unique.
 bool _itemsAreUnique(List items) => items.toSet().length == items.length;
 
-/// Checks whether all items in [items] are in [universal].
+/// Checks whether all items in `items` are in `universal`.
 bool _itemsExistInUniversal(List items, List universal) =>
     items.every((item) => universal.contains(item));
 
-/// Gives [k]th permutation in the ordered list of permutations of
-/// items taken from [items].
+/// Gives `k`th permutation in the ordered list of permutations of
+/// items taken from `items`.
 List _permutationWorker(BigInt k, List items) {
   int n = items.length;
   if (n <= 1) {
@@ -44,8 +44,8 @@ List _permutationWorker(BigInt k, List items) {
   }
 }
 
-/// Gives the index of [permutation] in the ordered list of permutations of
-/// items taken from [items].
+/// Gives the index of `permutation` in the ordered list of permutations of
+/// items taken from `items`.
 BigInt _inversePermutationWorker(List permutation, List items) {
   if (permutation.length == 1) return BigInt.zero;
   int n = items.length;
@@ -59,8 +59,8 @@ BigInt _inversePermutationWorker(List permutation, List items) {
       (group % BigInt.two == BigInt.zero ? biN - index - BigInt.one : index);
 }
 
-/// Gives [k]th combination in the ordered list of combinations of
-/// [r] items taken from [items].
+/// Gives `k`th combination in the ordered list of combinations of
+/// `r` items taken from `items`.
 List _combination(BigInt k, int r, List items) {
   if (r == 0) return [];
   int n = items.length, position = 0;
@@ -75,8 +75,8 @@ List _combination(BigInt k, int r, List items) {
   return [items[position.toInt()]]..addAll(_combination(k, r - 1, tail));
 }
 
-/// Gives the index of [combination] in the ordered list of combinations of
-/// items taken from [items].
+/// Gives the index of `combination` in the ordered list of combinations of
+/// items taken from `items`.
 BigInt _inverseCombination(List combination, List items) {
   BigInt helper(List combination, List items) {
     if (combination.isEmpty) return BigInt.zero;
@@ -94,8 +94,8 @@ BigInt _inverseCombination(List combination, List items) {
   return helper(_sortedArrangement(combination, items), items);
 }
 
-/// Gives [k]th selection in the ordered list of selections of
-/// [r] items taken from [items].
+/// Gives `k`th selection in the ordered list of selections of
+/// `r` items taken from `items`.
 List _selection(BigInt k, int r, List items) {
   int n = items.length, position = 0;
   BigInt d = _nCr(n + r - position - 2, r - 1);
@@ -104,16 +104,16 @@ List _selection(BigInt k, int r, List items) {
     position += 1;
     d = _nCr(n + r - position - 2, r - 1);
   }
-  if (r == 0)
+  if (r == 0) {
     return [];
-  else {
+  } else {
     List tail = items.sublist(position.toInt());
     return [items[position.toInt()]]..addAll(_selection(k, r - 1, tail));
   }
 }
 
-/// Gives the index of [selection] in the ordered list of selections of
-/// items taken from [items].
+/// Gives the index of `selection` in the ordered list of selections of
+/// items taken from `items`.
 BigInt _inverseSelection(List selection, List items) {
   BigInt helper(List selection, List items) {
     if (selection.isEmpty) return BigInt.zero;
@@ -132,16 +132,16 @@ BigInt _inverseSelection(List selection, List items) {
   return helper(_sortedArrangement(selection, items), items);
 }
 
-/// Gives [k]th permutation in the ordered list of permutations of
-/// [r] items taken from [items].
+/// Gives `k`th permutation in the ordered list of permutations of
+/// `r` items taken from `items`.
 List _permutation(BigInt k, int r, List items) {
   BigInt f = _fact(r), group = k ~/ f, item = k % f;
   List comb = _combination(group, r, items);
   return _permutationWorker(item, comb);
 }
 
-/// Gives the index of [permutation] in the ordered list of permutations of
-/// items taken from [items].
+/// Gives the index of `permutation` in the ordered list of permutations of
+/// items taken from `items`.
 BigInt _inversePermutation(List permutation, List items) {
   int r = permutation.length;
   if (r == 0) return BigInt.zero;
@@ -151,8 +151,8 @@ BigInt _inversePermutation(List permutation, List items) {
       _inversePermutationWorker(permutation, sortedPermutation);
 }
 
-/// Gives [k]th amalgam in the ordered list of amalgams of
-/// [r] items taken from [items].
+/// Gives `k`th amalgam in the ordered list of amalgams of
+/// `r` items taken from `items`.
 List _amalgam(BigInt k, int r, List items) {
   return List.generate(r.toInt(), (i) {
     BigInt p = BigInt.from(items.length).pow(r - i - 1), index = k ~/ p;
@@ -161,13 +161,15 @@ List _amalgam(BigInt k, int r, List items) {
   });
 }
 
-/// Gives the index of [amalgam] in the ordered list of amalgams of
-/// items taken from [items].
+/// Gives the index of `amalgam` in the ordered list of amalgams of
+/// items taken from `items`.
 BigInt _inverseAmalgam(List amalgam, List items) {
   int r = amalgam.length;
   var n = BigInt.from(items.length);
   var powers = List<BigInt>.filled(r, BigInt.one, growable: false);
-  for (int i = 1; i < powers.length; i++) powers[i] = powers[i - 1] * n;
+  for (int i = 1; i < powers.length; i++) {
+    powers[i] = powers[i - 1] * n;
+  }
   return List.generate(
       r,
       (position) =>
@@ -175,8 +177,8 @@ BigInt _inverseAmalgam(List amalgam, List items) {
           powers[r - position - 1]).fold(BigInt.zero, (a, b) => a + b);
 }
 
-/// Gives [k]th subset in the ordered list of subsets of
-/// items taken from [items].
+/// Gives `k`th subset in the ordered list of subsets of
+/// items taken from `items`.
 List _subset(BigInt k, List items) {
   k = _adjustedIndex(k, BigInt.one << items.length);
   List r = [];
@@ -188,8 +190,8 @@ List _subset(BigInt k, List items) {
   return r;
 }
 
-/// Gives the index of [subset] in the ordered list of subsets of
-/// items taken from [items].
+/// Gives the index of `subset` in the ordered list of subsets of
+/// items taken from `items`.
 BigInt _inverseSubset(List subset, List items) {
   BigInt helper(List subset, List items) {
     BigInt k = BigInt.zero, power = BigInt.one;
@@ -203,8 +205,8 @@ BigInt _inverseSubset(List subset, List items) {
   return helper(_sortedArrangement(subset.toSet().toList(), items), items);
 }
 
-/// Gives [k]th compound in the ordered list of compounds of
-/// items taken from [items].
+/// Gives `k`th compound in the ordered list of compounds of
+/// items taken from `items`.
 List _compound(BigInt k, List items) {
   int n = items.length, r;
   for (r = 0; r < n; r++) {
@@ -218,8 +220,8 @@ List _compound(BigInt k, List items) {
   return _permutation(k, r, items);
 }
 
-/// Gives the index of [compound] in the ordered list of compounds of
-/// items taken from [items].
+/// Gives the index of `compound` in the ordered list of compounds of
+/// items taken from `items`.
 BigInt _inverseCompound(List compound, List items) {
   BigInt k = (List.generate(compound.length, (r) => _nPr(items.length, r)))
       .fold(BigInt.zero, (a, b) => a + b);
@@ -237,7 +239,7 @@ BigInt _indexFromIntOrBigInt(Object k) {
   } else if (k is BigInt) {
     biK = k;
   } else {
-    throw Exception("Index must be an int or BigInt.");
+    throw Exception('Index must be an int or BigInt.');
   }
   return biK;
 }